文法フォームの使い方のモーダルをコンポーネント化した。
