+++ system/src/CodingForm.vue
@@ -167,27 +167,7 @@
       </div>
       <CodeRun ref="codeRunner" />
       <!--文法フォームを選んだときに出す情報モーダル-->
-      <div v-show="grammarFormInfoIsShow" class="grammarFormInfo">
-        <div class="grammarModal">
-          <div class="grammardModalInner">
-            <div class="modalUpperArea">
-              <h1>文法フォームの使い方</h1>
-              <p>
-                文法フォームを使ってコーディングし、要素の構成を理解しましょう。
-              </p>
-            </div>
-            <div class="modalButtonArea">
-              <button
-                class="css-button-rounded--sand"
-                @click="closeGrammarModal()"
-              >
-                戻る
-              </button>
-              <button class="css-button-rounded--green">ダウンロード</button>
-            </div>
-          </div>
-        </div>
-      </div>
+      <GrammarMmodal ref="gModal" />
     </div>
   </div>
 </template>
@@ -202,6 +182,7 @@ import { indentWithTab } from "@codemirror/commands";
 import Toggle from "@vueform/toggle";
 
 import CodeRun from "./components/CodeRun.vue";
+import GrammarMmodal from "./components/grammer_modal.vue";
 
 //関数定義、制御構造フォーム
 //フォームを追加する場合は、ここにコンポーネントを読み込む
@@ -217,6 +198,7 @@ export default {
     Codemirror,
     Toggle,
     CodeRun,
+    GrammarMmodal,
 
     funcForm,
     ifForm,
@@ -240,7 +222,6 @@ export default {
       nextIsDisabled: false,
       sampleIsShow: true,
       notificationIsShow: false,
-      grammarFormInfoIsShow: false,
       //フォームのタブの状態
       formTabSelected: 1,
       //関数定義フォーム、制御構造フォームの種類
@@ -465,10 +446,7 @@ export default {
       });
     },
     openGrammarModal: function () {
-      this.grammarFormInfoIsShow = true;
-    },
-    closeGrammarModal: function () {
-      this.grammarFormInfoIsShow = false;
+      this.$refs.gModal.openGrammarModal();
     },
   },
 };
@@ -743,43 +721,6 @@ div.tabGroup {
 div.tabContent {
   width: 100%;
   height: 500px;
-}
-
-/*文法フォームモーダルの設定*/
-div.grammarFormInfo {
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  position: fixed;
-  margin: 0;
-  background: #868686;
-  z-index: 999;
-}
-div.grammarModal {
-  width: 600px;
-  height: 500px;
-  margin-top: 200px;
-  margin-left: auto;
-  margin-right: auto;
-  background: #ffffff;
-}
-div.grammarModalInner {
-  width: 90%;
-  height: auto;
-  margin: 0 auto;
-}
-div.modalUpperArea {
-  width: 100%;
-  height: auto;
-  margin-top: 100px;
-  margin-left: auto;
-  margin-right: auto;
-}
-div.modalButtonArea {
-  width: 90%;
-  height: auto;
-  text-align: center;
 }
 </style>
 <style src="@vueform/toggle/themes/default.css"></style>
